Refactor the ground rendering

+++ src/game/assets/sprites.rs
@@ -22,16 +22,19 @@ impl SpriteData {
         let mut sprite_sheet = graphics::Image::new(ctx, "/sprites.png")?;
         sprite_sheet.set_filter(FilterMode::Nearest);
 
-        let sprite_sheet_size = [
-            sprite_sheet.width() as usize,
-            sprite_sheet.height() as usize,
-        ];
+        let sprite_sheet_size = [sprite_sheet.width() as f32, sprite_sheet.height() as f32];
 
         Ok(SpriteData {
             sprite_sheet,
-            digits: Digits::new(sprite_sheet_size, [655, 2], [9, 11], 1),
-            ground: Ground::new(sprite_sheet_size, [2, 54], [30, 12])?,
-            dino: DinoData::new(sprite_sheet_size, [848, 2], [44, 47], [1112, 19], [59, 30]),
+            digits: Digits::new(sprite_sheet_size, [655.0, 2.0], [9.0, 11.0], 1.0),
+            ground: Ground::new(sprite_sheet_size, [2.0, 54.0], [30.0, 12.0])?,
+            dino: DinoData::new(
+                sprite_sheet_size,
+                [848.0, 2.0],
+                [44.0, 47.0],
+                [1112.0, 19.0],
+                [59.0, 30.0],
+            ),
         })
     }
 }
